pyproject.toml
[project]
name = "mvp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.5",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.3",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
]
